requirements.txt
Django==5.0.1
asgiref==3.8.1
sqlparse==0.5.3
djangorestframework==3.14.0
django-filter==23.5
drf-yasg==1.21.7
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-redis==5.4.0
gunicorn==21.2.0
whitenoise==6.6.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
django-storages==1.14.2
boto3==1.34.34
pytest==8.0.2
pytest-django==4.8.0
coverage==7.3.2
selenium==4.15.2
factory-boy==3.3.0

# ML dependencies
mlflow==2.10.2
scikit-learn==1.4.0
pandas==2.2.0
numpy==1.26.3
transformers==4.37.2
torch==2.2.0
